Combine all spec tests into one file
